renamed my components and rendered an input field
