+++ src/components/Auto/CheckCard/CheckCard.jsx
